Chuyen De Phat Trien Ung Dung 2 - TDC - 2024 Tuan: 6 Ngay: 02/11/2024 De tai: Quan ly sieu thi Nhom: 2 Chau Nhat Tai Nguyen Quoc Luong Le Van Toan
Noi dung cong viec:
- [x] Edited SRS, sort by database (Chapter 2, 3)
- [x] Updated db scripts (Insert, Scripts, Scripts_Data, Added Query_LastRecord)
- [x] Coded frmLog, responsive frmLog, frmKhuyenMai
- [x] Added DataValidation.cs in BUS

/* Committed? */
- [x] Yes/No

/* Note */
- Check testcase for frmsNhatTai yet

+++ PTUD/CDUD - 2/files/src code/Nhom2 - PTUD - QLST/DAL/DAL_TaiKhoan.cs
@@ -87,7 +87,7 @@ namespace DAL
                          {
                              Quyen = tk.Quyen
                          }).FirstOrDefault();
-            return query.Quyen;
+            return (int)query.Quyen;
         }
 
         public void AddTK(DTO_TaiKhoan taiKhoan)
